feat: create rating saga added

+++ src/GoodReads.Application/Features/Ratings/Create/CreateRatingRequest.cs
@@ -5,7 +5,7 @@ using MediatR;
 namespace GoodReads.Application.Features.Ratings
 {
     public sealed record CreateRatingRequest(
-        decimal Score,
+        int Score,
         string Description,
         CreateReadingRequest Reading,
         Guid UserId,
